requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
faiss-cpu==1.8.0.post1
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-core==0.2.28
langchain-text-splitters==0.2.2
langsmith==0.1.98
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.32
tenacity==8.5.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
Flask==2.0.1
requests==2.25.1
faiss-cpu==1.7.2
numpy==1.21.0
PyMuPDF==1.18.14
